updated_urlbuilder/requirements.txt
cachetools==4.1.1
centaurminer==0.0.8
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
google-api-core==1.22.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.27.2
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
pandas-gbq==0.13.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
webdriver-manager==3.2.2
